adding unit test and small refactors

+++ api/src/main/java/hr/task/api/model/CompanyMessageLog.java
@@ -6,15 +6,15 @@ public class CompanyMessageLog extends MessageLog {
 	private static final String COST = "Cost";
 	private static final String PROFIT = "Profit";
 
-	public void updateRevenue(long revenue) {
+	public void updateRevenue(int revenue) {
 		super.update(REVENUE, revenue);
 	}
 
-	public void updateCost(long cost) {
+	public void updateCost(int cost) {
 		super.update(COST, cost);
 	}
 
-	public void updateProfit(long revenue, long cost) {
+	public void updateProfit(int revenue, int cost) {
 		super.update(PROFIT, revenue - cost);
 	}
